Tutorials first
I always think a tutorials first approach to documentation is much more inviting, because it becomes immediately clear to the user (a) what is solved and (b) how to use it. All of the extra details of what other things you can choose to do can come after the tutorial, but most people want to just look at the first tutorial, copy/paste, and run.

With that in mind, I think the first tutorial could use a bit more. I think a demonstration of surrogate optimization should appear in that first page, otherwise most people won't know it exists.